move main and init back
